requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
aiohttp==3.10.10
apscheduler==3.10.4
python-dotenv==1.0.1
